Use int instead of long in windows typedefs for stdtype typedefs. This interoperates better with other projects.

+++ inc/qcc/windows/platform_types.h
@@ -32,8 +32,8 @@ typedef signed char int8_t;           ///< 8-bit signed integer.
 typedef unsigned char uint8_t;        ///< 8-bit unsigned integer.
 typedef signed short int16_t;         ///< 16-bit signed integer.
 typedef unsigned short uint16_t;      ///< 16-bit unsigned integer.
-typedef signed long int32_t;          ///< 32-bit signed integer.
-typedef unsigned long uint32_t;       ///< 32-bit unsigned integer.
+typedef signed int int32_t;           ///< 32-bit signed integer.
+typedef unsigned int uint32_t;        ///< 32-bit unsigned integer.
 typedef signed long long int64_t;     ///< 64-bit signed integer.
 typedef unsigned long long uint64_t;  ///< 64-bit unsigned integer.
 #endif
